Lambda/renameFiles/requirements.txt
Pillow==6.1.0
urllib3==1.25.5